pyproject.toml
[project]
name = "mavia-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "amadeus>=12.0.0",
    "azure-storage-blob>=12.25.1",
    "black>=25.1.0",
    "blpapi-mcp==0.1.0",
    "browser-use>=0.1.24",
    "bs4>=0.0.2",
    "cohere>=5.14.2",
    "composio-langchain==0.7.7",
    "cryptocompare==0.7.6",
    "einops==0.8.1",
    "fastapi==0.115.6",
    "fastembed>=0.6.1",
    "flashrank[listwise]>=0.2.10",
    "flask>=3.1.0",
    "gdeltdoc==1.5.0",
    "google-auth==2.37.0",
    "google-search-results>=2.4.2",
    "isort>=6.0.1",
    "json-repair>=0.41.1",
    "langchain==0.3.20",
    "langchain-aws==0.2.13",
    "langchain-community==0.3.19",
    "langchain-core==0.3.43",
    "langchain-google-vertexai==2.0.14",
    "langchain-huggingface==0.1.2",
    "langchain-mcp-adapters>=0.0.7",
    "langchain-openai==0.3.7",
    "langchain-qdrant==0.2.0",
    "langchain-tavily==0.1.5",
    "langchain-text-splitters==0.3.6",
    "langchainhub==0.1.21",
    "langgraph==0.3.5",
    "langgraph-checkpoint==2.0.18",
    "langgraph-prebuilt==0.1.2",
    "langgraph-sdk==0.1.55",
    "langgraph-supervisor==0.0.9",
    "langsmith==0.3.13",
    "lingua-language-detector==1.4.0",
    "lxml>=5.3.2",
    "markdown>=3.8",
    "mcp-use>=1.2.8",
    "mcphub[all]>=0.1.9",
    "motor==3.6.0",
    "mysql-connector-python==9.2.0",
    "openpyxl>=3.1.5",
    "pillow>=11.2.1",
    "playwright>=1.51.0",
    "psutil==7.0.0",
    "psycopg2-binary==2.9.10",
    "pyasn1-modules==0.4.1",
    "pydantic==2.10.4",
    "pyjwt==2.10.1",
    "pymongo==4.9.1",
    "pymupdf>=1.25.5",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-playwright>=0.7.0",
    "python-binance==1.0.25",
    "python-dotenv==1.0.1",
    "python-twitter-v2==0.9.2",
    "sqlalchemy==2.0.35",
    "streamlit>=1.45.1",
    "tweepy>=4.15.0",
    "uvicorn==0.32.1",
]

[tool.uv.sources]
blpapi-mcp = { git = "https://github.com/djsamseng/bloomberg-mcp" }
